backend/requirements.txt
python-dotenv==1.0.0
blinker==1.6.2
boto3==1.28.68
botocore==1.31.68
click==8.1.7
colorama==0.4.6
distlib==0.3.7
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
platformdirs==3.11.0
PyJWT==2.8.0
python-dateutil==2.8.2
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.7
virtualenv==20.24.5
Werkzeug==2.3.7